Add basic cleanup script #21
