+++ src/types.js
@@ -1,0 +1,12 @@
+
+/**
+ * @typedef {Object} BuildInfo
+ * @property {string} id
+ * @property {Object} body
+ * @property {Date} date
+ * @property {string} ref
+ * @property {string} repository
+ * @property {string} status
+*/
+
+exports.types = {}
